+++ packages/flutter_spanned_controller/CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ToggleStateListener`, which is a generalization of `AttributeListener`
 
+### Fixed
+
+- `GestureRecognizers` on spans are now indexed with their attribute instead of the resolved attribute value,
+  because function object equality does not work with inline lambdas.
+
 ## [0.4.0-0] — 2020-12-22
 
 ### Removed